Implement comprehensive validation framework for BitNet.rs
- Added property-based tests for greedy argmax invariants and teacher forcing consistency in `test_greedy_invariants.py`.
- Introduced utility functions for JSONL file handling in `util.py`.
- Documented validation framework, including tokenizer parity, logit parity, and NLL parity in `VALIDATION.md` and `VALIDATION_QUICK_START.md`.
- Summarized validation improvements and implementation status in `validation-improvements-summary.md`.
- Established a baseline performance metrics file `perf_baseline.json`.
- Developed scripts for checking greedy argmax invariants and validating all components of the framework.
- Enhanced replay tools for debugging and comparing Rust and HuggingFace model outputs.
- Updated validation scripts to streamline the testing process and ensure robust performance checks.

+++ crates/bitnet-inference/src/engine.rs
@@ -221,7 +221,7 @@ impl InferenceEngine {
             
             // Capture logits if requested (after sampling to know chosen_id)
             if let Some(cb) = &config.logits_cb {
-                if step < config.logits_tap_steps {
+                if (step as usize) < config.logits_tap_steps {
                     let k = config.logits_topk.min(logits.len());
                     
                     // Use partial selection for efficiency on large vocabs
@@ -243,7 +243,7 @@ impl InferenceEngine {
                         .collect();
                     
                     // Pass topk and the chosen token
-                    (cb)(step, topk, next_token);
+                    (cb)(step as usize, topk, next_token);
                 }
             }
